DataFormat change! Using now same widths as CMSSW for variables

+++ DataFormats/interface/KMetadata.h
@@ -15,8 +15,8 @@ struct KProvenance
 /* RUN + LUMI METADATA */
 struct KLumiMetadata
 {
-	unsigned long nLumi;
-	unsigned long nRun;
+	unsigned int nLumi;
+	unsigned int nRun;
 	std::vector<std::string> hltNames;
 	std::vector<std::string> hltNamesMuons;
 	std::vector<std::string> discrTau;
@@ -35,10 +35,10 @@ struct KEventMetadata
 	unsigned long long bitsL1;
 	unsigned long long bitsHLT;
 	unsigned long bitsUserFlags;
-	unsigned long nEvent;
-	unsigned long nLumi;
-	unsigned long nRun;
-	unsigned long nBX;
+	unsigned int nEvent;
+	unsigned int nLumi;
+	unsigned int nRun;
+	int nBX;
 };
 
 // List of user flags
